Complete Phase 2 complexity refactoring: CreateMatchPage and CreateTournamentPage
This commit completes Phase 2 of the complexity refactoring plan by fully decomposing
the two remaining high-complexity React components.

## CreateMatchPage Refactoring (Complexity: 30→~8)

Reduced main component from 1233→426 LOC (65% reduction) by extracting:

**Custom Hook:**
- useMatchForm.ts - Form state management, session storage, map operations

**Helper Functions:**
- match-helpers.ts - Payload building, map notes saving, signup initiation

**Step Components:**
- GameSelectionStep.tsx - Game selection grid (Step 1)
- EventInfoStep.tsx - Event information form (Step 2)
- AnnouncementsStep.tsx - Announcement configuration (Step 3)
- MapConfigurationStep.tsx - Map selection and configuration (Step 4)

**Reusable Components:**
- EventImageUpload.tsx - Image upload with preview (shared with tournaments)
- SelectedMapsList.tsx - Map grid display with notes
- MapSelector.tsx - Mode-aware map selection UI
- MapCard.tsx - Individual map display card
- FlexibleMapCard.tsx - Map card with mode selection

## CreateTournamentPage Refactoring (Complexity: 28→~6)

Reduced main component from 701→261 LOC (63% reduction) by extracting:

**Custom Hook:**
- useTournamentForm.ts - Form state management, team operations

**Helper Functions:**
- tournament-helpers.ts - Payload building, team creation, signup transitions

**Step Components:**
- TournamentGameSelectionStep.tsx - Game selection (Step 1)
- TournamentEventInfoStep.tsx - Event info form (Step 2, reuses EventImageUpload)
- TournamentTeamSettingsStep.tsx - Team settings (Step 3)
- TournamentReviewStep.tsx - Review and confirmation (Step 4)
- TeamList.tsx - Pre-created teams display and management

## Key Benefits

- Each wizard step is independently testable
- State management isolated into custom hooks with session storage
- Shared EventImageUpload component promotes code reuse
- Clear separation of concerns: UI, state, business logic
- Significantly reduced cyclomatic complexity
- Improved maintainability and readability

## Phase 2 Status: ✅ ALL COMPLETED (9/9 functions)

Updated COMPLEXITY_REFACTORING.md to reflect completion of all Phase 2 items.
